Update patch set 1
Patch Set 1:

If that's the style you prefer, please remove the WIP and I'll merge it. I'm all for cleaning things up.

One could make the argument that you maybe want the bool, but as a template parameter (and DCHECKs). That would allow compiler optimizations at the price of space.

Patch-set: 1
Label: Verified=0