Update patch set 8
Patch Set 8:

> High-level question: when do you plan on doing the patching?

Sorry. The general idea is if we have a boot.oat that has an option indicating that patches are included (using agampe's new methods) then the image_space.cc code that currently finds an image will relocate on demand to create one, similar to how GenerateImage works if there is no image at all (similar in that it will lazily do it only when checksum's diff it already exists).

Similarly isDexOptNeededInternal and ClassLinker::CreateOatFileForDexLocation can relocate app .oat files, but that part is still a work in progress.

Patch-set: 8