Update patch set 4
Patch Set 4:

I specifically asked for the test case because it tracing keeps regressing.

I think its a good idea to add --trace support, but that isn't as critical as getting a sanity test in my opinion.

I don't support running --trace on all tests all the time for test-art, it is already slow enough.

Patch-set: 4