Update patch set 3
Patch Set 3: Code-Review+1

(1 comment)

I guess we should understand whether the FIXME in gen_common.cc:823 is really something that needs fixing or not (personally, I don't find it necessary). If not, then we need to remove the FIXME and leave the remaining part of the comment.

Patch-set: 3
Label: Code-Review=+1