Update patch set 1
Patch Set 1:

Nicolas: this CL is up for discussion of a problem I faced...

When the target is 64b, and maybe not PREFER_32, the nativetest directory is now called nativetest64. That was the case for me for both arm64 and x86-64 last night with ToT. It would be great if somebody could confirm that I'm not going crazy.

Patch-set: 1