Update patch set 1
Patch Set 1: Code-Review-1

Please wait and don't submit this.  The code is almost identical to the x86 code and we could find a way to merge them rather than maintaining two code bases.

It also will not compile on the mac due to the differences in ucontext on Mac OS X and Linux.

I would prefer that we get the 4x reverted x86 change in first and then see how we can meld the 64 bit support into it.

Patch-set: 1
Reviewer: Gerrit User 1036486 <1036486@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1