Update patch set 1
Patch Set 1:

(2 comments)

The best way to approach this is to have the x86_64 build compile fault_manager_x86.cc and use #if defined(__x86_64__) in the file to switch between the architectures.  There is already an #defined(__APPLE__) for the ucontext differences between mac os x and linux.

Patch-set: 1