Update patch set 1
Patch Set 1:

If I am reading the ARMv7 spec correctly, then orr sets (some) condition codes, and exactly the one we'd need here. Instead at the moment we often generate a cmp + b.ne, as the distance to the slow path is too large for a cbz. Maybe we should fix this?

Patch-set: 1
Label: Verified=0