Update patch set 1
Patch Set 1:

This looks good, however the current encoding for kThumb2OrrRRR does not seem to set the "S" field. Thus kOpOr is actually a ORR reg reg re instead of a ORRS reg reg reg (It I read the encoding table correctly :-)).

"F7.1.128
ORR (register)
Bitwise OR (register) performs a bitwise (inclusive) OR of a register value and an optionally-shifted register value,
and writes the result to the destination register. It can optionally update the condition flags based on the result.

where:
setflags = (S == ‘1’);
"

Patch-set: 1