Update patch set 1
Patch Set 1:

We might not have a valid PC or PC-to-Dex-PC mapping. In fact, currently we can't retrieve the fault PC (missing API). So we pretend the PC is 0.

Monitor inflation wants to record which method locked an object, and at which dex PC. It uses the standard CurrentMethodVisitor for that, which basically says: if you cannot get the dex PC, fail.

This CL changes that to: if you cannot find the dex PC, return kDexNoIndex (0xFFFFFFFF) and let the caller deal with it; and ignore kDexNoIndex in the monitor (pc is only used for contention logging).

Patch-set: 1