Update patch set 1
Patch Set 1:

What will the effect be on the target?  It will fail dex2oat with a fatal error but what will the contents of the dalvik-cache be?  I assume there will be nothing there and then another attempt will be made to compile it - infinite loop (but of course, it might work second time).

Patch-set: 1
Reviewer: Gerrit User 1036486 <1036486@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0