Update patch set 1
Patch Set 1:

Of course that is very troublesome. This is why the check was added to all the callers.

But when we're dying already and want to dump as much of the state as we can, the implicit check will make us abort and lose *all* info of the original crash.

Patch-set: 1