Update patch set 1
Patch Set 1:

I'd appreciate comments on the RegClassForAbsFP(). I don't know if load/store of a core reg is slower/faster that load/store of an fp reg. For example we may just choose the kFPReg all the time. And the choice may be different on arm and arm64.

See also https://android-review.googlesource.com/101914 .

Patch-set: 1
Label: Verified=0