Update patch set 1
Patch Set 1:

By user I meant a ART developer: there is utility function to get constant basing on ssa reg but it contains a DCHECK and if you want to get constant value for high part representing by ssa reg, utility function does not allow you to do this due to we did not mark this ssa reg as constant. This CL fixes this by marking it as constant.

Patch-set: 1