Update patch set 3
Patch Set 3:

(2 comments)

I'd like the choice of using a register vs a RIP load to be under the control of a flag. I think there's a bug in that RIP loads aren't being generated and that x86-64 has inherited the "start of code" logic from 32-bit that can't do RIP addressing.

Patch-set: 3