Update patch set 1
Patch Set 1:

I don't think I can get to this with a unit test, as the current QBE won't generate a call to this sequence.  It does run the unit tests for reverseBytes in 082-inline-execute in my code, which is how I noticed that it wasn't implemented.

Patch-set: 1
Reviewer: Gerrit User 1036869 <1036869@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0