Update patch set 1
Patch Set 1:

I'm not sure I follow. According to http://developer.android.com/reference/java/lang/Runtime.html#freeMemory() The definition of freeMemory is very much linked to totalMemory. It explicity says it Is how much memory can be allocated without extending the currently allocated heap. 

Regardless of that fact, this change is only intended to revert back what is currently broken. The first change done to FreeMemory was not actually needed in the first place, and it was followed by a succession of patches and it's still broken as it now gives inconsistent results with other functions. If we roll that change back, and add the test with what is assumed of those functions then you guys can work on the proper fix :)

Patch-set: 1