Update patch set 1
Patch Set 1: Code-Review-2

I'm concerned that OpRegImm doesn't handle kOpRev for ARM and MIPS and results in an abort. If we have a different contract for this virtual method on x86 to ARM and MIPS we're likely to get bugs. Perhaps what's missing here is an instruction canonical-izer and then assertions on what the contract of OpRemImm is. Given there's also no testing, I'd prefer not to take this change.

Patch-set: 1
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2