Update patch set 5
Patch Set 5:

If we have a bug in a switch statement on a byte then I have confidence anybody with an instruction set manual can go in and fix it. The levels of complexity added here makes maintenance harder. I'm prepared to do this investment but it should be shared across the code base. If we're being too lazy to do that let's go with a simple decode routine.

Patch-set: 5