Update patch set 1
Patch Set 1:

This is fine by me. I'm happy to go back to the state it was before. Although, we do need a CL that increments the version, right?

If there's a discussion to have, it's about compiler engineers, that do want the boot image to be re-generated because of some optimizations, or changes that make the boot image different but still compatible with existing compiled dex code.

It'd be great to have a way of knowing whether we need a version change. But I know the difficulty of having it.

Patch-set: 1
Label: Verified=0