Update patch set 3
Patch Set 3:

I did some investigations myself. I confirm that the size of boot.oat increases by 0.04% with this patch. I generated the oatdump of boot.oat with and without this patch. I grep-ped over these outputs for the occurrences of ldr, str, stp, ldp. This should give a naive/rough estimate of the effect of this patch on various instructions. I found that this patch reduces the occurrences of ldr, ldp, stp by 0.45% (3638), 2.69% (2608), 2.71% (2653) respectively. However, it increases the occurrences of str by 4.11% (6668). It also increases the occurrences of movz, movk by 5.72% and 0.66%, respectively. On one side, having less ldp/stp is good as these are more "expensive" than ldr/str. On the other side, an ldr/str is more likely to be placed inside a loop. Below I give an example of how the code is typically affected.

Without the patch:

  stp x24, x25, [sp, #40]
  str x30, [sp, #56]
  ...
  mvn w25, wzr
  cmp w20, w25
  ...
  ldp x24, x25, [sp, #72]
  ldr x30, [sp, #88]

With the patch:

  stp x24, x30, [sp, #32]
  ...
  mvn w2, wzr
  str w2, [sp, #24]
  cmp w20, w2
  ...
  ldp x24, x30, [sp, #64]

This is - I think - the most frequent pattern that emerges by doing the diff: reduction of number of ldp/stp, increase of number of str. I wonder whether we really need to store the constant back into the stack in the example above. My feeling is that overall the effect should be slightly beneficial, but it may be difficult to see it in benchmarks (AFAICS, the variance of running popular benchmarks tens of times is often ridiculous).

Patch-set: 3