Update patch set 1
Patch Set 1:

(4 comments)

This should really go to the arm/arm64 specific classes.

On arm, you can use ORRS+BMI to check (dstPos < 0 || srcPos < 0).

On arm64, you can use CCMP to reduce the number of branches. If it's properly reduced to a single branch, the slowpath doesn't need a special launchpad and the loop's back-branch can be conditional, saving an instruction on each iteration; though this would slow down the len_too_big path and we may not want that.

Patch-set: 1