Update patch set 2
Patch Set 2:

Thanks for your comments.

These are the numbers that Baptiste has found.
On two different processors, and from 4 to 3000 entries, the speedup is betwee 2.23 and 4.21x

For codesize, with 4, 10, 50, 100, 1000 & 3000 entries the size of the original version to the patched is:

1.3, 1.4, 1.5, 1.0, 0.86, 0.84

So between 4 and 100 entries (inclusive), the patched version is smaller.

Above 3000 there is another threshold where the code is not compiled and the interpreter is used instead.

I'll discuss with Baptiste what he thinks of lifting this to the MIR.

Patch-set: 2
Reviewer: Gerrit User 1037216 <1037216@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0