Update patch set 2
Patch Set 2:

The code size will certainly depend on the keys, i.e whether they are inexpensive and can be embedded in the CMP or not. Note that the keys are ordered, so we can use SUBS with the differences instead of CMP and further reduce the code size in some cases. (Counterexamples where the size would increase can be constructed but they are not very likely to show up in normal code.)

Patch-set: 2
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0