Update patch set 7
Patch Set 7:

> .oat files are not updated when compiler has changed. Maybe we
 > should add compiler as a dependency for .oat files?

no, we just reverted the change someone did to do that. it doesn't help on the device. the solution is to bump the oat version in oat.cc. This will force the host recompile as well as target.

Patch-set: 7
Reviewer: Gerrit User 1003723 <1003723@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0