Update patch set 2
Patch Set 2:

This raises - possibly - an outstanding issue: in AssembleLIR() we currently assume that methods are not larger than 1 MB. To be more explicit, if - during fixup - we find out that a branch jumps to a target which is more distant than 1 MB, then we just abort with LOG(FATAL). This may be fine, as - if I recall correctly - we skip compilation for large methods. I still have difficulties understanding how much of an issue this is in practice.

Patch-set: 2
Label: Code-Review=0