Update patch set 3
Patch Set 3:

I found this solution a bit over complicated and hard to follow. Maybe some more comments and rationale would help.

In particular, I don't really understand the need for the string_cache list.

Also it might be cheaper (and I think also cleaner and easier to understand) to just add another (canonical) mapping to the original table when we don't find one. This will need to be bounded to avoid attacks IMO. This is actually a mix between the solutions.

Patch-set: 3
Label: Verified=0