Update patch set 4
Patch Set 4:

To Andreas: OK, thanks for the reply. I confess I didn't investigate this in detail. In particular, I didn't check the actual limits for the dex code or whether the switch table payload - for example - would be "counted" as subject to this limit or not.

Patch-set: 4
Reviewer: Gerrit User 1037505 <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0