Update patch set 1
Patch Set 1: Code-Review-2

-2: This breaks SGET/SPUT.

The problem is that the fast path for SGET/SPUT needs two things: the class needs to be initialized and the type entry must be in the dex cache. When we do the INVOKE_STATIC and hit the class init trampoline, we initialize the class but we do not update the type entry in the dex cache. Subsequent SGET/SPUT still needs to do the check. See comments in https://android-review.googlesource.com/87694 .

The correct approach would be to separate the two things and keep two distinct flags. Then we could also improve the medium path for SGET/SPUT when only one of those flags is set.

Patch-set: 1
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2