Update patch set 3
Patch Set 3:

This one uses alias_info for HeapRef accesses. I decided to use 0 as the "ambiguous" value. This precludes optimizing accesses to the class field, but is for safety reasons, as it's the default value (so someone not calling AnnotateHeapRefAccess can't mess up the analysis).

Note that I believe we're broken on non-SMP systems: as we don't generate barriers, we can eliminate volatile loads and stores.

Patch-set: 3