Update patch set 3
Patch Set 3:

I had to add an alias field to the ShadowFrame. Because frame addresses are used as frame identifiers I need to keep a frame-likeish relation between them, but there are situations where one creates a ShadowFrame on the heap breaking this relation. This happens when we replaces a quick frame with a shadow frame and I solved this by storing the original quick frame address in the new shadow frame and use that address as frame identifier instead of the address of the shadow frame.

Patch-set: 3
Label: Verified=0