Update patch set 3
Patch Set 3:

I have implemented step-out by adding a new InterpreterWillBeExited event to the InstrumentationListener and sending that event from artQuickToInterpreterBridge before returning.
In response to that event the debugger will request a frame deoptimization by setting the deoptimization exception, thus we will be able to step into the upcall that was compiled originally.

Patch-set: 3