Update patch set 3
Patch Set 3: -Code-Review

Thanks for review. This is an optimization we ported from x86 dalvik code since divides take many cycles and are fully blocking, anything to not take divide path can lead to an overall win.
This is a work in progress since we need to port the rest of dalvik divide optimization. Will update the patch with the rest of algorithm done in dalvik.

Patch-set: 3
Reviewer: Gerrit User 1013850 <1013850@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0