Update patch set 3
Patch Set 3: Code-Review-1

The SP and PC can be in the list in A32 (ARM) instructions, but not in T32 (Thumb) instructions. Also the usage of instructions where the base register is part of the list is not recommended. Please see section: F7.1.212 STM (STMIA, STMEA) in ARM ARMv8. 

http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.genc010197a/index.html

Patch-set: 3