Update patch set 3
Patch Set 3: -Code-Review

That is correct r14 is a valid register in the reglist for STM(STMIA, STMEA). I see that the other variants of this instruction, STMIA use reglist[0, 15], though reglist[13] and reglist[15] must be 0.

I will leave Bill to decide on this. Probably the proper fix is to define a new variant that DCHECKs the reglist.

Patch-set: 3
Reviewer: Gerrit User 1030338 <1030338@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0