Update patch set 1
Patch Set 1:

I've given this some more thought, and am beginning to think that the attempt to do a LONG_TO_INT conversion notationally on 64-bit targets is just plain wrong.  I reused this code from the 32-bit world, where it made sense (because longs were always either in memory or a pair of 32-bit registers).  But, I think it was a mistake for 64-bit targets.

Consider the following dex code fragment:

LONG_TO_INT v0, v0
return v0

Now, suppose that we're using the same temp reg for v0 that is also used as ret0 on the target.  By just doing the narrowing via RegLocation bookkeeping updates, we could allow crap high bits to bleed into the return result.  If that return value was immediately used as an array index by the caller you'd have the same problem (and this CL would not catch it).

The short-term solution here, I think, is for our 64-bit targets to always do a zero-extend for LONG_TO_INT when the source operand is in register.  It's true that we'd end up with some useless zero extends - but I think the cost would be minimal.

I did a quick count of LONG_TO_INTs during the compilation of the entire boot class path: 957 total, 257 of which operated on the same vreg.  Only a subset of these would cause additional code to be generated (no change to vregs in memory, and the instruction count would not change for narrowings to a different register).  Further, I suspect the run-time cost of adding an occasional ALU op is effectively zero.

Thoughts?

Patch-set: 1