Update patch set 10
Patch Set 10:

(2 comments)

There were some bugs around the static initialisers.

One of them was reproducible by stopping in a static initialiser and placing a breakpoint in the caller (that called the static method and so implicitly initiated class initialisation), then issuing a continue command. Because the caller quick frame will invoke whatever function pointer the artQuickResolutionTrampoline returns to it, the need of stack deoptimization will not be recognised after the static method returned. We have solved this by forcing instrumentation on the static method, thus after the run of the static method the hooked instrumentation exit function will recognise the aforementioned need of stack deoptimization.

Patch-set: 10
Label: Verified=0