Update patch set 9
Patch Set 9: Code-Review+1

Vladimir, thanks for fixing the imm26 bug :-)

I think AOSP is currently configured to run as interpret-only. For our builds, we apply a patch to disable this. The patch is in platform/build (core/main.mk). We remove the lines,

  # Don't compile apps on eng builds to speed startup 		
  ADDITIONAL_BUILD_PROPERTIES += dalvik.vm.dex2oat-filter=interpret-only

Removing these two lines helps unveiling problems in the compilation of boot.oat. We do this also for builds that we use to run benchmarks. I am not sure whether this has to do with the problem you were experiencing yesterday. I can say that removing these two lines makes our builds unable to boot to GUI (boot.oat is miscompiled by the 64-bit host version of dex2oat). We solve the problem internally by applying https://android-review.googlesource.com/#/c/107401/ (note that this is an arm64-specific problem).

Patch-set: 9
Label: Code-Review=+1