Update patch set 1
Patch Set 1:

> > > Could you be more precise? What naming conflicts? And why have
 > > this
 > > > ginormous change, if the change in the optimizing compiler
 > would
 > > > likely be smaller?
 > >
 > > I understand this touches many lines for a renaming, but I
 > believe
 > > it is worth it. Two reasons for this patch:
 > > 1) It improves the current naming by having clearly prefixed
 > > register enums 'X', 'W', 'D', 'S'. (With the current scheme, what
 > > is a 'Register'? Is a 'Wregister', 'Dregister', or 'SRegister' a
 > > 'Register'?
 > > 2) It will make the opt compiler code much easier to write and
 > more
 > > readable.
 > > The opt compiler code generation will rely heavily on VIXL. To
 > > avoid having to prefix everywhere with 'vixl::', we will import
 > the
 > > vixl namespace in -arm64.cc files (only -  to not pollute other
 > > archs).
 > > VIXL exposes the 'Register' type (covering both X and W
 > registers),
 > > which is used everywhere in the VIXL helpers and will be used
 > > everywhere in the codegen.
 > 
 > I don't understand. In 1) you're saying Register is unclear but in
 > 2) that Register in VIXL is both X and W. Isn't what we have in ART
 > (Register is X or core) cleaner?

I think I understand the confusion from the comments.
Perhaps the most important point here is that vixl::Register and the current art::arm64::Register do *not* represent the same thing, and we are *not* attempting to replace art::arm64::Register. We are simply trying to avoid a naming conflict.
The Register *class* exposed by VIXL will be used heavily in the code generation, in a different way than art::arm64::Register is and still will be used. The renaming proposed in this patch would ease writing the arm64 opt compiler backend, while in our opinion not being detrimental - but instead useful - to the art::arm64::<?>Register enums and existing code.

A little more details about the vixl::Register class.
VIXL exposes the Register and FPRegister *classes* (both subclasses of CPURegister). They are not simply enum, but enforce register type and width coherency. It is simply more convenient to use for code generation (we use it everywhere), in particular because it is the interface provided by VIXL. (Otherwise I'm sure we could have worked out something based on art::ManagedRegister, but hey we already have VIXL so let's avoid some work!)
When I said that a vixl::Register allows both to cover X and W register, I did not mean it aliases the two, but that it could represent the two. So in VIXL, 'x0.Is(w0)' will evaluate to false, but 'x0.W() == w0' evaluates to false, etc.

Patch-set: 1