Update patch set 1
Patch Set 1:

aapcs64 define the fp behavior as : 
$fp = $x29
previrouse_fp = *($fp)
lr = *($fp + 8)

DS-5 streamline is a sample based profile tool. It will try to unwind the stack with the above information in runtime.

In this patch, we implement fp register as aapcs64 spec. So that streamline can get the call path.

A sample use case will be : both A() B() will invoke C(). C() is time consuming, but it is already quite optimal. We want to optimize the caller A() or B() to avoid the method call to C() if possible. By examining the call path in stream line, we can know which function calls C() heavily and consumes a lot of time. Without the ability to unwind the stack, we can only know the self executing time of each method.

This option should usually be used with --include-debug-symbols enabled at the same time. So that the profiling tool can get both call path and method names.

BTW, this feature cannot be implemented on thumb2, as aapcs does not define the fp behavior. And different tool chains implement different fp behavior, some of the implementation even relies on the frame info to unwind the stack, which is not loaded to the memory in runtime. So the profile tool can not get call path on thumb2.

See: http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.dui0482q/ric1362355425191.html

Patch-set: 1