Update patch set 1
Patch Set 1: Code-Review+1

Andreas, to reproduce to bug we need a complex condition, specifically: allocation of object which has finalize() method declaration should pass but when adding it to finalizer queue some small object is created and OOM should happen at this moment. We can try to create some reproducer but it will not 100% fail, is it ok? Also reproducibility can be depended on heap size...

Patch-set: 1
Label: Code-Review=+1