Update patch set 14
Patch Set 14:

> Whe have fixed some bugs in the latest patch sets.
 > 
 > There were a bug where we step out from the static initializer and
 > interpreter was forced instead of instrumentation, thus the stack
 > deoptimization was performed at the end of the interpreter after
 > the called method ran in interpreter. We have fixed this and added
 > a call to Dbg::IsForcedInterpreterNeededForUpcall in
 > instrumentation::PopInstrumentationStackFrame. So, when we step out
 > from the static initializer we will require instrumentation on the
 > called method and in the instrumentation exit hook we will
 > recognise the need of stack deoptimization by the previously
 > mentioned call.
 > 
 > Also, the reflection bug has been solved, by an extra call in
 > ArtMethod::Invoke, just as you assumed.
 > 
 > We have made minor optimizations by reordering the if statements
 > and adding an extra one: the Dbg::IsForced* methods will return
 > false if we are in interpreter only mode. To measure the
 > optimizations a very simple test of calling the same method
 > numerous times was performed and from that I got around 5-6 percent
 > of improvement in non-interpreter mode.
 > 
 > Regarding the benchmarks: Is there a canonical place somewhere in
 > the Android source tree where we can get these benchmarks? When you
 > saw the 20% increase, were you running ART with the -Xint option? I
 > would like to reproduce your results to make it sure the regression
 > has been solved.

Thanks for the comments, it's clearer now. 

About the benchmarks, no they are not in the Android source tree. There are many different benchmarks in the field that can help you. As I said, I use CaffeineMark and Dhrystone to measure performance impact. In our case here, I believe you need an app doing lots of invoke to measure this change because it mainly apply to invoking method.

I didn't use -Xint mode yet because it'd be really slow. I'll do it as last step to verify. To measure impact on interpreter, I forced the benchmark to not be compiled using "-Xcompiler-option --compiler-filter=interpret-only" command line arguments. Therefore benchmark's code runs with interpreter while system's code keeps running with compiled code. Note the -Xcompiler-option is only seen by the compiler so it's only used when the application needs to be compiled which happens at install time. As I run benchmark from the command-line in my case, I ensure to always remove the DEX from the dalvik-cache before running it.

Patch-set: 14