Update patch set 3
Patch Set 3:

> Side question:   Does it make sense for x86_64 code generator to
 > derive from the x86 one?  A lot of the code seems similar.

My experience is that, while it's a nice thing to have, it's too much in the way of improving both backends (optimizations we'll do on x86_64 and not on x86 and vice versa). I sometimes think I'm over-engineering the shared code_generator.cc code already.

Patch-set: 3
Label: Verified=0