Update patch set 1
Patch Set 1:

Agreed, it would be worth finding out because you may have seen an issue with it.

Anyway, regarding kMirOpNullCheck, as per the specification I wrote, there is a single use of object register in vA. Therefore, the null check should be on use 0. At least this is my understanding so it would be great if you could confirm. :)

Patch-set: 1