Update patch set 1
Patch Set 1:

> Can an equivalent effect be had by just declaring id_ as id_:10 ?

I am not very familiar with C++ bitfields, but I don't think so.
The Arm64ManagedRegister uses the remaining bits to encode the register type. It expects those bits to be initialized to 0 by the ManagedRegister, so that we can detect if a ManagedRegister is constructed and improperly cast to an Arm64ManagedRegister. Also, C++ bitfields would not guarantee that the bits are preserved across static_casts.

Patch-set: 1
Reviewer: Gerrit User 1052304 <1052304@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0