Update patch set 3
Patch Set 3:

I cannot edit my comment but what I meant about updated memory location is that an optimizing compiler may not write to the VR stack location if it does not need to, especially for constants. Thus the memory load might be invalid data and thus the constant case seems like it should be handled.

Patch-set: 3