Update patch set 1
Patch Set 1:

It requires fewer changes than I thought it would.

I am also curious to know if using a template would be cleaner for slow paths.
One thing to be careful of: vixl::Labels should *not* be copied or moved. This will be required to handle automatic handling of distant branch targets and veneers insertion.
So for example since GrowableArray uses memcpy when growing, we can't template block_labels_ on the label class.

I'm starting to rebase.

Patch-set: 1