Update patch set 3
Patch Set 3:

I'd like to think about this one a little.   It may no longer be an issue, but what you're running into is the short end of an old intended trade-off.  Back in the trace JIT days (when we had no promoted registers, very little type info and even less time to do analysis) IGETs and SGETs usually involved a flush of the loaded value to a vreg in the frame.  For the vast majority of our supported devices, we had a much more rich set of load/store instructions targeting core registers than floating point registers.  So, it was a win to steer IGETs and SGETs into core registers even if we knew they were float.  Given the soft-float calling convention and frequent case of vreg copying, the cost of the occasional core->float reg copy was much less than the cost of having to expand floating point loads and stores into multi-instruction sequences because we were out of displacement range.

I can see this is a bigger problem for x86 - and much has changed since the trace JIT days so the old trade-offs may no longer apply.    Let me think about it - a more general solution may be appropriate now.

Patch-set: 3
Label: Verified=0