Update patch set 2
Patch Set 2: -Code-Review

> then also passing an argument that is only used by the assembler

Forgive me for being thick here, but what argument are we talking about? From what I see it's the same set of args passed to the C compiler plus -E, did I miss something?

At any rate, what is the recommended path to resolve this? Right now none of the ART code is successfully being cached, which is quite an annoyance when doing clean builds. Do you want us to make a change in build/ such that '-Wa,--noexecstack' is only passed to the linker?

Patch-set: 2