Update patch set 5
Patch Set 5:

We do NOT have patches for gcc to address this, or patches to get gcc optimizer behavior back to what you saw with 4.8.

A minority view in the mips gcc team is that gcc's de facto definition of strict-aliasing rules are too fragile, for cases like this of loads and stores inlined from separate modules.  But any fix of those rules would involve the worldwide gcc community; not likely to change soon.  Our bug report complaining about the outcome in this case, was rejected.

The gcc change which exposed this problem, adjusted inlining pragmatics for all machines, increasing the size of the affected routine by 3x but reducing the size of other routines, for a net improvement.  Somehow this stopped the unifying optimization for the store and reload, on some target variant arches, including 2 mips variants.  We do not have a gcc change to redo the inlining balance.

I have not heard how the changed inlining managed to sometimes block the store/load unification. That is unfortunate and hopefully fixable. We do not understand that or have a gcc change to improve that.

Patch-set: 5
Reviewer: Gerrit User 1010669 <1010669@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0