Update patch set 1
Patch Set 1: Code-Review-1

Thanks for taking an early look!

This is WIP (Work-in-Progress) so I didn't get a chance to clean it up yet, definitely do not intend to merge it in this state.

Uploaded only at Andreas's request since my WIP patchoat change for --compile-pic is using this, and he was looking at changing it to use an oat header field instead.

Patch-set: 1
Reviewer: Gerrit User 1021471 <1021471@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1