Update patch set 1
Patch Set 1:

This solution is a bit of hack, but would I venture that this code will almost never be generated.  I wrote this mostly to have test 083-compiler-regressions pass, since more methods of this test are getting compiled with the optimized compiler, as support for Dex instructions expands.

I guess a better solution would be to detect whether a stack frame is too large to be accessed with an offset fitting in an immediate value prior to generating code for the corresponding method -- and if so, to allocate an additional stack slot to save the IP register in Thumb2Assembler::StoreToOffset.

Another solution would be to refuse to compile methods with too large stack frames.

Patch-set: 1
Label: Verified=0