Update patch set 1
Patch Set 1:

Notice this is more a safety measure for the future than a real fix.

It can be shown indeed that the current version of Art is unable to generate a sequence of instructions matching the erratum conditions. Future optimisations, however, may make a hit possible. This CL addresses this issue at the assembler level in the Quick Compiler.

Patch-set: 1
Reviewer: Gerrit User 1037505 <1037505@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0