Update patch set 1
Patch Set 1:

Agree with Vladimir Marko. If the index is negative int, we should signed extend that in GenArrayGet()/GenArrayPut() not regard it as 64-bit reg with high part cleared. The result is same "ArrayIndexOutOfBoundsException", but signed extend is the right thing we should do here.

Patch-set: 1
Reviewer: Gerrit User 1042649 <1042649@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0