Update patch set 1
Patch Set 1:

> > We generally dislike system property code in ART.
 > 
 > Right, just make it a runtime option and pass the option in based
 > on the property from AndroidRuntime.cpp
 > 
 > but then again, why not just use
 > 
 > adb shell setprop debug.db.uid 1
 > 
 > to have debuggerd catch the process to let you attach gdb?

Dmitry, correct me if I'm wrong, but this is about large-scale automated testing, where you want to see the dumps far after the fact? So gdb is a bit out of the question.

Instead of reading the system property in ART and another runtime option, why not add a bit to the debug_flags being passed in from the framework...

Patch-set: 1