Update patch set 2
Patch Set 2:

> > > I don't follow this. If the intent is to avoid dealing with
 > bugs
 > > > during bring up I worry the consequence will be missing bugs
 > when
 > > > things are brought up. We shouldn't be doing it.
 > >
 > > My reading of this is that he is trying to do the minimal build
 > > possible to improve turn around time. If he is only building
 > libart
 > > to run to save time, he doesn't want to build libartd just to run
 > > dex2oatd.
 > 
 > Indeed, thanks Brian.
 > 
 > So this is for enabling
 > $ export BUILD_ART_HOST_DEBUG=false
 > $ export BUILD_ART_HOST_NDEBUG=true
 > $ m -j32 test-art-host-run-test-ndebug
 > 
 > Unfortunately, there are too many debug dependencies for
 > test-art-host, that this CL is not enough,

Yes, but this is conflating lots of different things in a way that wasn't intended. We should have a way to run debug and ndebug tests, so that we can ensure we have coverage of both. This change doesn't address that, its doing a point hack which means we lose dex2oatd compilation on performance sensitive tests. The only reason there is support for libart currently is for performance sensitive runs. I'm not sure I get why we have regression tests on performance, but anyway, don't just hack away do a proper change where we have debug and ndebug in the make files, rationalize what we do with the performance sensitive tests, make sure we're defaulted over to ndebug... The optimizing compiler is worse than most of ART in using DCHECK almost exclusively so we don't get any invariant checking when not running the debug version. In terms of running tests faster I'd suggest more parallelism over losing assert coverage.

Patch-set: 2