Update patch set 2
Patch Set 2:

Andreas, I have a question. Inside the optimization there is a check for ISA in order to apply it. Should I make it be x86 isa? I could also make it the runtime ISA but then I need to consider the case that the transformation is not applied (like in mips case). What are your thoughts?

Patch-set: 2