Update patch set 4
Patch Set 4:

> Unfortunately, I found more functions in art/compiler with more
 > than 1728 bytes of stack frames. Would you prefer a simple limit as
 > in this CL, or a different one in each subdirectory of art?

Are there many functions in the compiler that are larger? Ideally we should be fixing these not to require such a large amount of stack space. Could you provide a list?

It'd be ok with me to have a limit on the disassembler and the compiler, with the default for the runtime.

Patch-set: 4
Reviewer: Gerrit User 1010118 <1010118@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0