Update patch set 1
Patch Set 1: Code-Review-1

My understanding is that is incorrect bytecode. At least I see that there should be into-to-float conversion between 0001 and 0004. The probblem will become more evident if in 0000 you will use the constant fitting int size but not representable in float. I hope that this code cannot be obtained by dx tool, so my proposal to reject such code from compilation in verifier.

Patch-set: 1