Update patch set 1
Patch Set 1:

This code cannot be produced by the dx tool, you are right. That is not to say however that this code cannot be generated at all: see the provided smali test. Are you able to run it? Does it produce the correct value?

It is legal in the verifier for an SSA reg to contain both a float and an integer value if that value is constant, since we must use the same bytecode to load both int and float constants. I don't see how we can reject this code during verification.

The verifier accepts the code, and so it is arguable that the code must be compilable, and the produced code must do the correct thing.

Patch-set: 1