Update patch set 1
Patch Set 1: -Code-Review

Hi All,

I did not say that this code cannot be generated or executed or it always produces invalid result. My point was about how often we can meet this code. Is it usual pattern produced by dx tool or it is very rare case while code is generated manually by some utilities. My point was that accepting this code for compilation needs to revisit all usages of location.fp field and make changes. The question is whether it is desirable or simply we can reject compilation of these methods if they are really rare. As I've understood Bill's note this is code which can easily meet. So it means that we should to find a way how to handle it. Quick look up shows me that we have some utility functions in gen_invoke.cc like CallRuntimeHelperRegLocation which actually a kind of invocation (like this patch deals with) and it relies on location fp attribute to detect what register to use. So I guess we need to review all usages of the location.fp and fix all of them if it is needed. I'm just interesting whether you have already checked all other passes or just fixed the case with invocation?

Patch-set: 1