Update patch set 1
Patch Set 1:

(2 comments)

Now about the patch itself. I've put my comment. The patch looks good in general to me. x86 implementation will need to be adjusted as well. 

However I think that it makes sense to generalize this solution making the common code everything except the mapping algorithm of the next argument to physical register.

Does it make sense at all? And does it make sense to do in terms of this patch or separate one?

Patch-set: 1