Update patch set 1
Patch Set 1:

> > > > Should we compile tests with kEverything instead?
 > > >
 > > > We could, but this is not the default behavior (that is the one
 > > we
 > > > ship), and I'd prefer just testing the default behavior.
 > >
 > > I think this is too brittle. Who actually checks that we're
 > testing
 > > the compiler (besides the breakage in the build bot...)?
 > 
 > ? There's no difference between how the buildbot runs and how your
 > machine runs. You would also have seen the breakage. The reason for
 > the breakage is the $opt$ marker on the method name, which checks
 > that the method is indeed compiled.

I was trying to say: who checks continuously that the tests are being run the way they are supposed to?

I'm surprised about the $opt marker, which is this check (didn't see that you added that), but personally I think it's a hack. Are you gonna go back and add the marker to *every* run-test, as we supposedly want to test all things in the compiled fashion?

Patch-set: 1