Update patch set 1
Patch Set 1:

I looked at the Barrier use in BackgroundMethodSamplingProfiler::RunProfilerThread() in profiler.cc.

We may be able to replace it with a (non-reusable, one-shot, countdown()/wait()-style) latch in that case with the following approach:

- We create a new Barrier each iteration of the while loop to avoid a reuse.

- We pass the uninitialized barrier to ThreadList::RunCheckpointOnRunnableThreads() and let it initialize the barrier count right after the for loop (where we know the final value of 'count') but before releasing Locks::thread_suspend_count_lock_ (which a runnable thread would block on before executing the checkpoint routine, as in Thread::RunCheckpointFunction()).

- We need to separate the construction and the count initialization of (Init()) to make it easy to share pointers to the Barrier between the checkpoint closure (eg SampleCheckpoint), ThreadList::RunCheckpointOnRunnableThreads(), and the main routine (eg RunProfilerThread()).

If this works and there's no other cases of barrier reuse, then we might be able to turn (or rename) it into a latch.

Patch-set: 1
Label: Code-Review=0